+++ resources/views/components/table/columns/toggle.blade.php
@@ -1,5 +1,8 @@
+@props(['item', 'column'])
+
 @php
-    $value = (bool) $column->getValue($item);
+    $field = $column->key;
+    $value = (bool) $item->{$field};
 @endphp
 
 <td
@@ -25,11 +28,34 @@
         ])
         style="text-align: {{ $column->align }}"
 >
-    <livewire:laravel-simple-datatables::toggle-switch
-            :model="$item"
-            field="{{ $column->key }}"
-            :key="'toggle-'.$item->id"
-            align="{{ $column->align }}"
-            textColor="{{ $column->color }}"
-    />
+    <div
+            x-data="{ active: {{ $value ? 'true' : 'false' }} }"
+            x-init="itemId = '{{ $item->id }}'; field = '{{ $field }}';"
+            @update-toggle-state.window="if (event.detail.itemId === '{{ $item->id }}' && event.detail.field === '{{ $field }}') { active = event.detail.value; }"
+            class="flex items-center"
+    >
+        <label class="relative h-6 w-12">
+            <input
+                    type="checkbox"
+                    x-model="active"
+                    wire:click="toggleValue('{{ $item->id }}', '{{ $field }}')"
+                    class="custom_switch peer absolute z-10 h-full w-full cursor-pointer opacity-0"
+                    id="switch_{{ $item->id }}_{{ $field }}"
+            />
+            <span
+                    for="switch_{{ $item->id }}_{{ $field }}"
+                @class([
+                    'dark:bg-dark dark:before:bg-white-dark block h-full rounded-full bg-[#ebedf2] before:absolute before:bottom-1 before:left-1 before:h-4 before:w-4 before:rounded-full before:bg-white before:transition-all before:duration-300 peer-checked:before:left-7 dark:peer-checked:before:bg-white',
+                    match ($column->textColor) {
+                        'secondary' => 'peer-checked:bg-secondary',
+                        'success' => 'peer-checked:bg-success',
+                        'danger' => 'peer-checked:bg-danger',
+                        'warning' => 'peer-checked:bg-warning',
+                        'info' => 'peer-checked:bg-info',
+                        default => 'peer-checked:bg-primary',
+                    },
+                ])
+            ></span>
+        </label>
+    </div>
 </td>